# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "糖尿病本体 × 患者事实库 融合查询",
    "description": "糖尿病本体与患者事实库融合查询 API。技术验证用途，不是医疗器械，不构成医疗建议。无鉴权、无限流，仅应部署在可信内网或本机。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patients": {
      "get": {
        "summary": "检索并分页列出患者",
        "description": "SQL 侧检索。分页与筛选都在这里做完 —— SPARQL 只处理收敛后的小集合。\n\n【用于】不知道患者编号，或需要按 ICD-10、事实来源、风险档位筛选和分页。\n【不要用于】在图里扫患者；已知单个编号问病情（直接用患者族对应工具）。",
        "operationId": "list_patients_patients_get",
        "parameters": [
          {
            "name": "icd10",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Icd10"
            },
            "description": "按 ICD-10 外部诊断编码前缀筛选，例如 E11。可不传。",
            "example": "E11"
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ehr-legacy / derived / demo-cohort",
              "title": "Origin",
              "enum": [
                "ehr-legacy",
                "derived",
                "demo-cohort"
              ]
            },
            "description": "按事实来源筛选：ehr-legacy 为真实上游数据，derived 为规则推导数据，demo-cohort 为演示队列。可不传。"
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario"
            },
            "description": "按演示场景编号筛选，例如 S02；主要用于调试演示队列。可不传。"
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "High / Moderate / Low / Insufficient-Evidence",
              "title": "Tier",
              "enum": [
                "High",
                "Moderate",
                "Low",
                "Insufficient-Evidence"
              ]
            },
            "description": "按规则式风险档位筛选；档位是有序枚举，不是概率或分数。可不传。"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page",
              "minimum": 1
            },
            "description": "页码，从 1 开始。",
            "example": 1
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size",
              "minimum": 1,
              "maximum": 200
            },
            "description": "每页记录数，范围 1–200。",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Patients Patients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "x-test-example-title": "检索 E11 患者并返回前两条"
      }
    },
    "/patients/{pid}/care-chain": {
      "get": {
        "summary": "获取患者完整照护链",
        "operationId": "care_chain_patients__pid__care_chain_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            },
            "description": "患者业务编号，例如 P90002。",
            "example": "P90002"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Care Chain Patients  Pid  Care Chain Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "患者、模板或资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "description": "单个患者的照护链：就诊 → 检验 → 诊断 → 用药的时间线与关联关系。\n\n【用于】「这个患者经历了什么」「什么时候查的」「在用什么药」。\n【不要用于】阈值判定（用「获取阈值判定与逐字出处」）、风险档位（用「获取规则式定性风险分层」）、用药安全信号（用「获取用药安全信号」）。\n【前置】需要患者业务编号；不知道编号先用「检索并分页列出患者」。",
        "x-test-example-title": "查询 P90002 的照护链"
      }
    },
    "/patients/{pid}/assessment": {
      "get": {
        "summary": "获取阈值判定与逐字出处",
        "description": "阈值判定 + 所用阈值区间 + 逐字出处 + sha256。\n\n⚠️ 诊断级切点 `confirmationRequired=true` 时，单次检验落在区间内 **≠ 确诊**；`verificationStatus=\"Provisional\"` 和 `caveat` 必须原样保留，不得升级成确诊。\n【不要用于】风险档位（用「获取规则式定性风险分层」）、时间线（用「获取患者完整照护链」）。\n【前置】需要患者业务编号；不知道编号先用「检索并分页列出患者」。",
        "operationId": "assessment_patients__pid__assessment_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            },
            "description": "患者业务编号，例如 P90002。",
            "example": "P90002"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Assessment Patients  Pid  Assessment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "患者、模板或资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "x-test-example-title": "查询单次 A1C 7.4% 的阈值判定"
      }
    },
    "/patients/{pid}/risk": {
      "get": {
        "summary": "获取规则式定性风险分层",
        "description": "风险分层。⚠️ 规则式定性分层，不含概率、不含发生率、不含时间窗；`tier` 是有序枚举，不是分数，也不是模型置信度。\n\n`Insufficient-Evidence` 既不是 Low，也不是「无风险」，更不是系统故障 —— 必须连同 `insufficientReason` 一起说明。`countedInTier=false` 的因子可以列出，但要讲明它不参与档位计算及其出处缺口。\n【不要用于】阈值判定或诊断依据（用「获取阈值判定与逐字出处」）、用药安全（用「获取用药安全信号」）。\n【前置】需要患者业务编号；不知道编号先用「检索并分页列出患者」。",
        "operationId": "risk_patients__pid__risk_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pid"
            },
            "description": "患者业务编号，例如 P90020。",
            "example": "P90020"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Risk Patients  Pid  Risk Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "患者、模板或资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "x-test-example-title": "查询有多个可改变风险因子的患者"
      }
    },
...
    "/graph/node": {
      "get": {
        "summary": "获取图节点邻接摘要",
        "description": "节点邻接摘要：类型（标出哪些是推理机推的）、所在图、出边/入边 + 计数。\n\n⚠️ 拒绝 `urn:dmo:data` 里的反例夹具，并说明原因 —— 悄悄过滤等于静默少返。\n【用于】拿到 IRI 之后的第一跳：看清类型和有哪些谓词可走，再决定下一步。\n【不要用于】展开某个具体关系的全部邻居（用「展开图节点一跳邻居」）、查上位/下位类（用「查询概念的上位或下位层次」）。\n【前置】必须已有准确 IRI —— 先用「将表面形式解析为准确 IRI」，**不要自己拼 IRI**。",
        "operationId": "graph_node_graph_node_get",
        "parameters": [
          {
            "name": "iri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iri"
            },
            "description": "要查看的完整节点 IRI；必须先通过 /graph/concepts 获得准确 IRI，不接受自然语言术语。",
            "example": "https://example.org/dmo/id/threshold/A1C-DIABETES"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Graph Node Graph Node Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "业务拒绝或请求不满足执行条件。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "GraphDB 暂时不可用。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Graph exploration"
        ],
        "x-test-example-title": "查看 A1C 糖尿病阈值节点"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Error": {
        "type": "object",
        "description": "标准错误响应。400 也可能表示有意的业务拒绝。",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "description": "错误或业务拒绝原因。部分端点可能返回结构化对象。",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "hint": {
            "type": "string",
            "description": "可选的排障建议。"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://124.223.18.44:8100",
      "description": "远程调试服务"
    }
  ],
  "security": [],
...
}